+++ AdventOfCode2021/Day13/day13.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <string>
 
-int pointsAfter1Fold(const std::string& fileName);
+size_t pointsAfter1Fold(const std::string& fileName);
 void pointsAfterEveryFold(const std::string& fileName);
